Improvements to Pathfinding
  - Each PathNode now holds a reference to its parent node.
  - Separated Open and Close sets.
  - The reconstructPath method walks backward from the goal node using parent pointers. Once the goal is reached, it builds the complete path and reverses the order so that it starts from the starting position and ends at the target.

+++ Common/src/main/java/dk/sdu/sem/commonsystem/Timer.java
@@ -11,7 +11,7 @@ public class Timer {
 
 	/// @param dt Delta time in seconds
 	/// @return True if the timer has elapsed, false otherwise
-	public boolean update(float dt) {
+	public boolean update(double dt) {
 		if (dt < 0) { return false; }
 
 		timer += dt;
